requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic-settings>=2.5.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
pdf2image>=1.17.0
Pillow>=10.4.0
python-multipart>=0.0.12

# LLM providers (install the one you need, or all)
anthropic>=0.34.0
google-genai>=1.0.0
openai>=1.40.0
